requirements.txt
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
black==23.11.0
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.104.1
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
Mako==1.3.0
MarkupSafe==2.1.3
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==4.1.0
pluggy==1.3.0
protobuf==4.21.12
pydantic==2.5.2
pydantic_core==2.14.5
pytest==7.4.3
python-decouple==3.8
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
uvicorn==0.24.0.post1
